+++ app/api/routes/v1/payments/transactions.py
@@ -1,34 +1,12 @@
 from typing import Annotated
-from uuid import UUID
 
-from fastapi import APIRouter, Depends, HTTPException, Query, Request, status
+from fastapi import APIRouter, Depends, status
 
 from core.dto.v1.payments import TransactionListItemDTO, TransactionListQueryDTO
-from core.foundation.dependencies import P24ServiceDep, PostgresSession
+from core.foundation.dependencies import AuthorizedTenantId, P24ServiceDep, PostgresSession
 from core.foundation.http.responses import PaginatedResponse
-from core.foundation.tenant_guard import resolve_and_authorize_tenant
 
 router = APIRouter()
-
-
-async def resolve_transactions_tenant_id(
-    request: Request,
-    session: PostgresSession,
-    tenant_public_id: str | None = Query(default=None),
-    tenant_id: str | None = Query(default=None),
-) -> UUID:
-    resolved_public_id = tenant_public_id or tenant_id
-    if resolved_public_id is None:
-        raise HTTPException(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail="tenant_public_id or tenant_id is required",
-        )
-
-    return await resolve_and_authorize_tenant(
-        tenant_public_id=resolved_public_id,
-        request=request,
-        session=session,
-    )
 
 
 @router.get(
@@ -37,7 +15,7 @@ async def resolve_transactions_tenant_id(
     response_model=PaginatedResponse[TransactionListItemDTO],
 )
 async def list_transactions(
-    tenant_id: Annotated[UUID, Depends(resolve_transactions_tenant_id)],
+    tenant_id: AuthorizedTenantId,
     session: PostgresSession,
     p24_service: P24ServiceDep,
     query: Annotated[TransactionListQueryDTO, Depends()],
